Combine MainView, Navigator as Routes
