+++ projects/newbies-layout/src/App.js
@@ -1,16 +1,15 @@
 import React, { Component } from 'react';
 import { BrowserRouter } from 'react-router-dom';
 import Layout from './layout/Layout';
-import Navigator from './Navigator';
-import MainView from './MainView';
+import Routes from './Routes';
 
 class App extends Component {
   render() {
     return (
       <BrowserRouter>
         <Layout
-          side={<Navigator />}
-          main={<MainView />}
+          side={<Routes.Navigator />}
+          main={<Routes.MainView />}
         />
       </BrowserRouter>
     );
